Moved code to do with types into resolve.go
